refactor(standard-version): split calver.rs into parse, bump submodules
Closes #211
